alexeyrnd-BJS2-48375. Mice are very nice. main() moved to clas MiceMain
